CovidData.jl

Some functions to download and combine various datasets related to Covid.

This was created for use with CovidSEIR.jl and CovidRt.jl.

CovidData.covidjhudata — Method

covidjhudata()

Downloads most recent JHU CSSE data on covid cases, deaths, and recoveries.

Returns a DataFrame

CovidData.statedata — Method

statedata(;filename="covidstates.Rda", policies=:dates, fillmissingmobility=false)

Loads data on US state case numbers, recoveries, deaths, policies, and movement. This dataset combines

JHU data on cases, deaths, & recoveries

NYT data on cases & deaths

Google mobility reports

Policy data from Raifman et al

Testing and hospitalizations from Covid Tracking Project

Operations of businesses and hourly workers from homebase

The downloading and combination of datasets is done in R.

Options:

policies in [:dates, :indicators] The downloaded policy date contains the date each policy went into effect. Set to :indicators if you want time-varying indicator variables for whether each policy was in place.

fillmissingmobility = false should missing values from Google mobility reports be filled forward and backward in time? If true, then prior to the start of data reporting, we set google mobility reports to 0, and after the most recent date report, we set them to the last non-missing value.

Returns a DataFrame